+++ Assets/_Scripts/Characters/ICharacterHealth.cs
@@ -1,6 +1,6 @@
 public interface ICharacterHealth
 {
-    public void TakeDamage(int amount);
+    public void TakeDamage(float amount);
     public void Die();
     public float GetCurrentHealth();
     public float GetMaxHealth();
